Move DFF to sequential
